Keep the farm DUTs and gitlab runners synchronized.
This component deals with three actors,

  1. A GitLab server instance.
  2. A local path to a gitlab-runner TOML configuration.
  3. A registry of known farm DUTs and their metadata.

And then ensures the server instance and the farm configuration are in
agreement about activated runners and their metadata.

Gitlab runner configuration has been moved to the permanent area,
since reconfiguring this is rather costly.

A "shim" registration service is here, completely BS for now, but it
serves its purpose.