add a machine_registration script
The script is meant to be run on the test machines, and will collect
all the information about it that is needed to register it in the
machine registration service (MaRS).

Use the 'register' action to generate the logs

Use the 'cachedb' action to pre-dowload the pciid database. Use the -a
parameter to reference the file.

Use the 'check' action to compare the current configuration with the
one stored in the MaRS.

References: #3