Consolidate NFS into boots
The latest update to 3.13 broke the NFS server. The,

capsh --print | grep -Eq "^Current: = .*,?${1}(,|$)"

Check for `cap_sys_admin` works when running privileged alpine:3.12
containers, but does not when running with alpine:3.13 containers.

I rolled the whole NFS server into boots and pinned the version for
future sanity.

Also get rid of the component-farm, and the associated madness with
submodules.