sergent_hartman: Use local treq, not global treq
The treq instance is dependency-injected into all these API classes,
in this case I missed a search and replace transitioning from the
global reactor-based treq to my injected one during a past
refactoring.

There is no behaviour change in this patch, but the consistency is
very important to avoid WTF-moments in the future.