salad: transition from a microservice to library design
SALAD as a separate microservice makes little sense as it has to remain
available as long as any job in the infra is running, and thus cannot
be updated independently. On the other hand, working as a microservice
makes it complicated to implement any kind of console streaming.

The general design is that SALAD is now a many-to-many multiplexer,
where serial ports's inputs/outputs are forwarded from/to a socket,
based on the machine ID they are connected to. The machine's ID and
state is inferred by snooping on the serial connection.

The new design is also easier to shutdown, as there is only *1* thread
that handles both the hot-(un)plug of new serial devices, rather than
having one thread per serial port. This was made possible by using
POSIX's select() on all the serial ports. This also simplified made it
easy to send bytes from multiple places in the codebase, as there
cannot be any sort of concurency.

The tests have disapeared in the process, but will be added again
later, when the design proves itself!