pdu-gateway: transition from a microservice to a library design
There is little point in having a network service for talking to PDUs,
as they are already network-enabled, and can be spoken to concurently.

The only case where this could become a problem would be when a new PDU
to an infra that would require a new driver to be written. This would
force a restart of the infra. This is however unlikely to happen, and
can easily be worked around by using the raw SNMP driver.