Add an executor service
This service is meant to allow "triggers" to start a job on the infra,
based on a job file which describes everything about the job:

 - target machine's tags
 - boot configuration (kernel, initramfs, kernel cmdline)
 - timeouts
 - success criteria
 - ...

The service abstracts the job deployment, power sequencing, and log
collection.

Interacting with this service should be done using executor/client.py,
which will queue the job, output in real time to stdout the logs from
the machine, forward stdin to the test machine for interactive
sessions, and set the exit code based on the job's status (pass, warn,
fail, complete, incomplete).