salad: close the serial console or the session when reading failed
By making sure to close the fd when we read 0 bytes from a "selected"
fd, we make sure the serial console / session that triggered it will
not be picked up for the next select call.

For serial consoles, this has the effect of freeing the serial device
name, so hotplugging the device again picks the same name.