artifact-reaper: New component
This is responsible for fetching the output of a job run, whatever
that might be. It works by mounting a volume over the output location
of the test-suite, which runs in a test container. Once the main test
job is complete, the post_container mounts the same volume, archives
everything, and sends it back to the job original job context.

This last step is important to Mesa Gitlab, since there, and directory
named "results/" in the PWD will get uploaded to the Gitlab
coordinator for, finally, inspection by users.

I elected to keep the job results in the local minio simply for
redundancy.