Force a rebuild of the containers
The deployment situation is unfortunate here. Because we had failed
pipelines in the past due to various issues, code that has landed
since on master has not triggered the container publish stage, and
code that has landed since has not modified the right files to
retrigger the past failed or skipped containers.

This left us in a state where everything was at random versions. Force
a rebuild of all containers to get back up-to-date. The mistake was
letting something land on master when the CI was broken, that's
because parts of the CI only run on the master branch, which I think
is a mistake, but not sure how best to deal with it right now.