gitlab-ci/pip: limit the execution to changes done to setup.cfg
This is a quick fix for the pipeline failure we are experiencing.

The proper solution is to stop depending on pypi for our infra, and
instead set up the version as something like:

    # pip3 install git+https://gitlab.freedesktop.org/mupuf/valve-infra@37e20d46251c6ad7068bda32bedf37c091060e57#egg=gfxinfo-mupufpkg&subdirectory=gfxinfo

This would enable us to depend on any git version, and we could push releases
manually after sufficient testing.