SALAD: convert from a library to a service
Having SALAD as a library was a good way to figure out the interface
that we would actually need in the executor, and it is now time again
to split it back into its own service.

The major difference from previous iterations is that rather than being
session-based, it is now simply doing one thing: routing from/to a
socket to/from a serial port. The routing decision is done by snooping
on the serial port, in search of "SALAD.machine_id" messages.

To let users know which port is associated to which machine ID, a REST
interface can be used to list the known machines:

GET /api/v1/machine

{
    "machines": {
        "18:67:b0:c3:84:7f": {
            "has_client": false
            "id": "18:67:b0:c3:84:7f"
            "tcp_port": 53813
        },
        "b4:2e:99:f0:76:c5":
            "has_client": false
            "id": "b4:2e:99:f0:76:c5"
            "tcp_port|: 58847
        },
        "d8:cb:8a:59:70:74":
            "has_client": false
            "id": "d8:cb:8a:59:70:74"
            "tcp_port": 56743
        },
    }
}

If wanting to query the information for a specific machine, and/or if
the machine of interest is not yet a known, you may instantiate it by
performing the following GET request:

GET /api/v1/machine/<machine_id>

{
     "id": "<machine_id>",
     "tcp_port": $port,
     "has_client": false
}

Finally, only one consumer per machine is possible, since we want to
enable full-duplex communication. So whenever a serial port will get
associated to a machine, any write to client socket connected to the
associated tcp_port will be forwarded to the serial port, and vice
versa.