executor: pull in gitlab-sync
This is needed in order to be able to pause runners when they become
busy, thus preventing them from getting picked up by GitLab to execute
tasks when they are actually busy doing something else.

This should have no functional differences with the previous code, but
I have taken the liberty to abstract the two classes with one. This
makes it simple to interface with it in mars.py.