entrypoint: Remove obsolete reference to private.env
For installed sites, this is now handled by the ci-installer's
environment installation. The farm name, gitlab registration and
access tokens, are all passed to the container in which we run as
environment variables, so the entrypoint may safely assume these are
set. If they are not set, the infrastructure can not feasibly start.