entrypoint: Fix bug in minio password generation
The completed set included the `-` character, which when present at
the start of the password, means to pass a CLI argument to whatever
context MINIO_ROOT_PASSWORD expands into. Highly unpleasant to track
down since it changes on each run.

Be more careful and only permit alpha numeric characters in the
password.