gfxinfo: Refactor VulkanInfo to handle construction exceptions cleanly.
It's not nice to suppress exceptions in a constructor, it leaves the
client unsure what the state of the return object is.

In these cases, a static constructor that properly returns the
exceptional status to the caller is preferred. In this way, the class
can take what it needs to serve its purpose, thus simplifying it.