Add a virtual PDU
Quick and dirty "PDU" for use with the virtual infra. A basic protocol
that emulates a PDU with on/off state management.

The guests log their output to a file in the CWD (where the disks and
other metadata live). This should help a bit when reasoning why things
runs were incomplete or failed. I didn't find a way to get the boot
ROM messages in the log yet, and that's where many INCOMPLETES are
coming from, due to the PXE client mysteriously dropping to a boot:
prompt rather than netbooting. Frequency of this happening has been
anywhere from 3 - 20% on the runs in my virtual farm!