executor/job: match console patterns using search instead of match
re.match can be very annoying, at it forces us to match whole lines
which may or may not include some \r characters. Rather than forcing
users to write regular expressions for the whole line, allow them to
when they want, and default to not doing that otherwise. We thus keep
all the power but just add convenience.

Fixes: c5f2d611881 (gitlab-trigger: use hang_detection to reboot th...)