job_templates: Interactive job
Commit this because I've made it several times now and don't want to
do it again. Rather handy for quick, interactive, tests. Linking this
up to the command line would be nice, i.e., removing the need to
hardcoded the machine id in the target yaml (it would instead be
rendered by the CLI)