Revert "entrypoint: Factor minio setup to its own service"
This reverts commit 8648bff3740880a66f20c8df75260e9d0dc32be6.

The buckets were being set, but the artifacts were not being
downloaded to the expected place. The "minio" alias had been moved
erroneously by the referenced commit, so the mcli cp was copying the
artifacts to $CWD/minio/... rather than the remote storage host!

Not found during testing, because these files were never removed for
me.