job templates: improve the timeouts
The register job does not need any console-related timeouts, as the
only communication that happens is pretty much when it has done its
work.

The bootloop job can be used to make sure the boot time of the machine
is reasonable by using the "first_console_activity", but there is
little point in a console_activity timeout when they would get close to
the overall timeout anyway. Let's also bump the overall timeout by a
minute to give time to the disk-formating process which can sometimes
take a while.