valvetraces, gfxinfo: more pycodestyle (formerly pep8) fixes
The one change of behaviour here would be in the regex literal in
valvetraces, which was a str, and tried using \d without
escaping. That's a bug which the code linter caught for us. Woohoo!

Add a pycodestyle (formerly pep8) checker to all pip stages to catch
regressions in the CI for pip packages.