executor: let the client download the job_bucket before its deletion
When the job is over, we want to let the client download the results
of the job execution back, so it may do further processing. However,
we want to make sure the bucket is deleted after the job terminates.

To do so, we introduce more job consoles states that will allow the
console to indicate that the end of the session has been hit, then we
let the execute_job() start the infra-teardown process which consists
of:

- Create credentials to the job bucket for the client to use
- Add the bucket's credentials to the console state
- Change the job console's state to TEAR_DOWN which triggers the
  end-of-session message which includes the bucket's credentials, and
  shutdown the writes to the client's socket
- Wait for the client to handle the message, call mcli mirror, then
  close the connection... which changes the job console's state to OVER

Finally, since we don't want to wait endlessly on the client to close
the connection, and we don't want to have a fixed timeout that may not
be suitable to all jobs, we make use the newly-introduced
infra_teardown timeout to get a configurable timeout.