executor/client: allow sending files as input for the job's execution
Building on the client's interface change, it is now possible to send
multiple files as part of the job submission process.

This is done by adding 2 parameters (--job-id, and --share-directory),
having the client tarball a user-specified folder, the executor
creating a new bucket and extracting the tarball there, then exposing
it to the job via previously-introduced job template variables.

Finally, the job bucket gets deleted at the end of the execution. More
work on this side will come in further commits.