executor/client: allow sharing MinIO privileges with the job
When jobs need to access sensitive data, the current way of sharing it
is to upload it to a MinIO bucket, then share the credentials as part
of the job description. This is not ideal, as these credentials can be
leaked and re-used later by jobs that were not supposed to have access
to such data.

To reduce the chances of credentials leakage, the client can now ask
the executor to tie these elevated access rights to the job's MinIO
user which is tied to the job's lifetime and the machine's IP address.
This way, when the job is done executing, the credentials expire.

Rather than explicitly sharing buckets/files, we instead ask the
executor to add the user to the groups specified by the client. To
prevent privilege escalation, we also ask the client to provide the
credentials for a MinIO user that is already in the groups that the
user is requesting. This way, the creation of the policy is left to
the user, which allows for full flexibility in access control.