executor: make the job bucket name more unique
It seems like the bucket existence code is not foolproof, which leads
to having too jobs using the same bucket... Oops!

By adding the machine id in the bucket name, we should get rid of the
issue, while we investigate what is going on with the bucket creation
code.

Additionally, we now make sure the bucket name is according to MinIO's
rules.