salad: Add a TCP network console server
While working on Vivian, it became clear that UDS were not the best
interface for "virtual" serial connections. They are specific to the
kernel they were created on, so sharing from multiple VMs is a
blocker. However, TCP sockets have none of these limitations, and QEMU
makes it very easy to back a serial device with a TCP socket.

Extending SALAD to listen for new "serial connections" via TCP
requests makes testing with virtual DUTs a piece of cake. Since they
supercede domain sockets, get rid of the old codepaths while we're at
it.