executor: Factor runner removal script to its own tool.
Does not make sense having these commands as part of the
executor. Since it's likely we'll want all sorts of scripts in the
future, add a new directory in valve-infra.

The runner removal sub-command was broken in executor anyway, since a
past commit remove the GITLAB_API context variable (and there's no
tests to catch that!). There's still no tests, but the tool was needed
earlier today.