vivian: Gracefully shutdown vpdu
It turns out that when launching a Python program from an
non-interactive shell, as is the case in Vivian, it's no longer
possible to send SIGINT to the child program. When launched from your
shell, everything works perfectly well. That is,

$ python vpdu.py &
$ pid=$!
$ kill -INT $pid

Works as you would expect, Cpython receives the interrupts, massages
it into a KeyboardInterrupt, and all is well. Not the case when the
above is executed via a shell-script!

One approach to tackling this is to install our own signal handler,
but signal handlers bring their own trickery. You can't use any
non-reentrant functions in such handlers, which rules out Python's
logging framework, for instance.

Instead, provide a flag to ask the server to shutdown over its TCP
socket.