valvetraces: refactor the report upload/junit generation under one roof
Both need to perform parsing, and could do with sharing an IR.

This is a huge commit that changes everything. I tried to separate it
into a lot of commits, but the series would be super long with many
intermediate states that would need to be extensively validated.

Given that the tool is currently only focused on reporting whether a
trace executed or not and no frames are missing, we don't have a good
base that does the expected checks (frame validation, checking the
execution history, ...). This makes splitting the commit a little
silly to split, as the clear change in objective makes bisection
impossible.

Still, dear reviewer, sorry about this!