valvetraces: correctly tag the jobs as pre/post-merge
Currently, we are planning on having developers push their branches
to a tree for testing, so we can't just check if the gitlab pipeline
is associated to a merge request. We thus also need to check that
the project we are testing can be considered upstream, and if the
branch is a part of the expected new run.