valvetraces/Trace: do not abort execution on error
Currently, when a trace fails to replay, we immediately abort the
execution of exec.sh due to "set -e" at the top of the script.

This means that the exit code and execution time never get captured,
which is not very good.

Just remove the 'set -e' flag to fix this issue.