vivian: don't use ovmf code/vars separately when efi booting
This seems to 'fix' the issue with booting the hard disk in efi, when
efi vars are handled separately it seems like the boot order is such
that booting the hard disk is much later in the order (after PXE and efi
shell, etc). At least in my testing on Alpine Linux, the default order
is CDROMS and then hard drives, so the gateway just boots like you'd
expect with this change.