ansible: Bundle chrony configuration
There was something fishy going on with the n.arch.pool.ntp.org
pool servers relating to the firewall. Remove the Arch specific pool
servers and use the global ntp.org ones, which works fine with our
existing firewall config.

While we're here, move all the Chrony configuration in the ansible
directory, rather than patching whatever is installed by Arch. The
later seeming a more brittle approach in general, since if the Arch
package updates and changes say the pool configuration, then the
firewall configuration would be out-of-date and we wouldn't know
it. In this case, I prefer the breakage be our fault than Arch's. In
another view, network time is a server we explicitly provide to the
infra, so it's better we're explicit about how we configure this
provided service.

The following configurations are uncommented from the default example3
configuration as well,

    -e '/^! leapsectz right\/UTC/s/^! //' \
    -e '/^! makestep 1.0 3/s/^! //' \
    -e '/^! rtcsync/s/^! //' \

These were kept from Arch's package.