ansible: add group_vars/gateway/default development var
This variable will be used by roles/tasks to perform config/actions
related to gateway development.

I went this route because:

- having a tag for this basically means duplicating some tasks, or
  setting a variable that roles can use to run things conditionally...
  so basically a slightly more complicated version of what I did here

- I tried using a separate playbook that basically extends gateway.yml,
  but again the only straight forward ways I could find to do things
  without duplicating roles was to set vars... so a more complicated
  version of what I did here

There are some downsides to doing this, the main one being I don't know
of a way to "undo" the configuration on a gateway once it is applied, at
least not without making new tasks that roll back changes (maintaining
those sounds terrible because they would have to change in sync with any
changes to the roles they undo).

The ci-installer's installer script can be modified to call
ansible-playbook with the development env set (controlled by a new
installer cmdline flag), so we can have 'development' ISOs quite easily.