valvetraces: only use the last 100 runs to check for output stability
With this change applied, we check that all the frames seen in the past
100 runs have been seen at least 20 times. The subtle thing to take
notice is that the 20 instances of a frame do not have to come from the
last 100 runs.

The result is that we filter out older instabilities, which means that
if an output was unstable a year ago, it cannot be a reason to accept
changes in rendering without a user checking the output.