ansible/*/handlers: only restart services if systemd is running
systemd may not be init, e.g. in cases where ansible is run as a part of
"{docker,podman} build".

Currently(?) testing "ansible_service_mgr" seems to be sufficient for
determining if systemd is running or not. In my testing, it was False
even if systemd is installed but not running/init.

If this changes in the future, testing that pid 1 is "init" might be
enough... ("ps -p1 -o comm=" or something)