executor: drop the MaRS service
This service is already an implementation detail of the executor, and
the usefulness of having a separate service just to store the
equivalent of a couple of kB of configuration is questionable.
Additionally, using sqlite to store what could essentially be a
configuration file makes backups harder to diff.

This commit drops MaRS, and replaces the database with a simple YAML
file which gets reloaded after every edit from a user or ourselves.