ipxe bootloader: Continuously retry boots
The chain command has curious semantics,

"Command status
Success	The image was successfully downloaded and booted, and the image subsequently returned successfully
Failure	The image was not successfully downloaded and booted, or the image subsequently returned with an error
It is unusual for this command to return successfully, since if the boot is successful then control will not return to iPXE."

Notice in particular that success does not make a lot of sense in this
case, and there are clearly conditions for the failure case that we
should simply keep retrying, to avoid needing manual intervention for
things like a misconfigured file location on the server.

Tested the recovery behaviour with QEMU, missing files cause the
bootloader to continuously loop trying again. Once the server was
reconfigured, the machine booted as expected.