Unfortunate side-effects of the container firewall
Unfortunately, the new valve-infra container with a firewall is going
to be a leaky abstraction. Its configuration will have to depend on
what network resources boot2container needs access to. For now, that's
only `backblazeb2.com', but this can change at any moment based on the
configuration of the volume storage made in the ipxe-boot-server.

So, even though we technically need this domain name whitelisted, it's
not going to help, because b2c resolves the address before the
container starts, which means when the container finally starts and
enables the default set of firewall blocks, the background volume sync
will stop working, and it will never be able to unblock itself, since
it doesn't use the resolver running in the started container. It could
do this by changing the host resolv.conf after the container started,
but this is not a workable design, it's far too brittle and error
prone.

Given this line of reasoning, it seems the firewalling should be
getting done at the b2c layer, not inside the guest container to avoid
this entire class of issues. We could potentially do something nice
using the podman networking infrastructure, or failing that,
configuring the packet filtering tables in the ramdisk based on the
ipxe-boot-server "secrets".

We could then change the firewall configuration using server-side
configuration in the boot server, depending on the container we choose
to run at a given farm.

Squash to firewall disable commit