ssh: Fix broken key files configuration
Because the pragma was given twice, the public keys we installed in
/etc/authorized_keys were not being used for authentication, they were
essentially ignored.

While I'm here, print the command in the Makefile because it's nice to
know, and get rid of the local network deescalation, since we should
be baking all developer a gateway keys into the container image now.