ansible: Support deployments to Vivian
Also tidy up the connection handling configuration. It was really
unfortunate putting the local connection info in the playbook, since
it meant very hard to debug behaviour when trying to override the
playbook connection type in the inventory (it would always fallback to
local), but this behaviour was not the case with the plain old ansible
command.

Be explicit about it!