ansible: Support optional environment files in the podman services
Due to the behaviour of run-time bind-mounting, it is not always
convenient to assume the perm mount's config.env will exist. In such
cases, do not fail the service start.

It's questionable what happens when the mirroring changes the perm
configuration, ideally the services should restart, but I don't think
systemd has such a feature for monitoring disk changes.