fdo_proxy_registry: Isolates REGISTRY_PROXY_REMOTEURL setting
REGISTRY_PROXY_REMOTEURL is meant only for the proxying registry. With
it being set in the global environment file, it adversely affected the
other registry service we host: the local registry.

Any jobs trying to push to this local registry would actually, by
proxy, try getting pushed to the fd.o registry. That operation would
return an Unauthorized access message, returned by the registry
protocol as HTTP 504. All of this was highly confusing to isolate.

The solution is to keep this environment in the container's unit,
rather than as a global service setting.