network_setup: Fix another $PRIVATE/private typo.
The gift that keeps on giving! In this case, the problem surfaced as
the private interface never being in the UP state (obviously). This
was missed because when I noticed it during testing, I simply set it
to UP manually and forgot to come back and check the root cause...