docs, build: Consolidate our build instructions & documentation.
Everything started getting a bit messy in the recent churn we've gone
through with the service architecture over the last months.

Consolidate all the docs in the top-level README.
Consolidate all build instructions into a top-level Makefile
Extend and clarify the documentation as it stands today.

Also make a start on the CI, the other services / containers should be
ported to build from the Makefile for extra developer clarity.