ansible, executor: Fix default PXELINUX configuration
The quoting screwed us over. The single quotes persisted into the
pxelinux configuration file, which when netbooted would cause the
virtual machine to insta-exit. This is because b2c doesn't pause or
show an error if the container it's asked to pull has an invalid
name. In this case, the container name was very broken due to the
spurious quotes.

It certainly looked innocent enough!