vivian: fix the boot sequence after the MaRS rewrite
These fixes should have gone into the MaRS rewrite series, but since
I unknowningly bypassed most of it when testing, I did not realize I
broke it completely.

The difficulty with this commit is that editing mars_db.yaml with
interactive editors work, but using sed/mv/... replaces the inode
which either fails when done from inside the container (device busy),
or simply does not impact the container... The solution is thus to
stop the service, edit the file, then start the service again.

Also, we need to do it twice, as we first need to unexpose the
current runners that may have come from the previous run, before
re-setting the mars config file.