ansible: Fix broken when conditions
The handler logic was working by accident, and when I tried running
this from within a container, handlers were mistakenly thinking
systemd is running. There's a difference of behaviour in the
environment of podman build and podman run from Ansible's
perspective. The former causes Ansible's reads from /proc to return a
fallback value of COMMAND\nansible-playboo (sic), the later falls back
to this method which makes ansible_service_mgr take on systemd as its
value,

  @staticmethod
  def is_systemd_managed_offline(module):
      # tools must be installed
      if module.get_bin_path('systemctl'):
	  # check if /sbin/init is a symlink to systemd
	  # on SUSE, /sbin/init may be missing if systemd-sysvinit package is not installed.
	  if os.path.islink('/sbin/init') and os.path.basename(os.readlink('/sbin/init')) == 'systemd':
	      return True
      return False

Hence, conditioning the handlers `when: ansible_service_mgr ==
"systemd"` is wrong, since it can be true, but still not usable for
system changes. It was "working" by accident within the podman build
environment, because of a separate issue, if you take a look at the
facts database in the build environment, you find,

        "service_mgr": "ansible-playboo",

Which is != systemd (and hence the handlers correctly don't run), but
what the heck is "ansible-playboo"? It turns out if you try to use
psutils within an environment that doesn't have permission to read
from /proc, you get "COMMAND\nansible-playboo" (sic, truncation) pack
from ps -p 1, which the service_mgr module helpfully tails for you
returning the strange "ansible-playboo", and things Just Work, but
this was definitely not the original intent.

So, stop Ansible's helpful massaging and return the PID 1 name as
a string to figure out whether we're in an environment which allows us
to change the system state with systemd. This feature gives us the
ability to provision running systems with changes in a safe way.