valvetraces: fix the generation of the jobtimeline
Until now, we have been providing the job_metadata field in the wrong
place, which means that the website never created a good timeline for
us... leading to tons and tons of duplicate timelines...

Simon and I reworked the job timeline to be more than just a dict(),
and instead have manadatory fields. This commit implements support for
it in valvetraces.