ipxe-boot-server: throw error if genfsimg fails
I don't think it makes much sense to continue on with the make process
if genfsimg fails for some reasons, e.g. the following sha256sum would
fail, and so on, and make would report "success" even when it didn't
make an iso. This causes make to return an error.