valvetraces: create our own entrypoint
Valvetraces depends on more files than just the trace execution
folders, making documentation and deployment of the tool harder than
it should.

By introducing our own entrypoint, we can take care of generating all
the files we need, and we can simplify our interface which will be
documented in an upcoming commit.