ci: Add a recipe for the base container
This will be used to more quickly run CI checks for ansible atop a
slow moving base. The ansible package is changed to ansible-core,
since the former has about 700MB of stuff we don't use! We do need
community.general for pacman though.

Using ci-templates was not possible for this, so this commit hacks
together a similar approach as workaround for now.

The version from Arch regressed in behaviour compared to the previous
version we were using from bullseye-slim.