vivian: keep mars_db across runs
After a discussion with Charlie, we agreed that it may be beneficial
to keep mars_db across runs, allowing us to test reboots.

This commit will make it so that mars_db.yaml is untouched, unless the
file is currently missing.

If the initial behaviour that every boot would be starting fresh is
wanted, then deleting the disk between runs will be necessary.