ansible: Parameterise the executor package location
During local builds (and CI builds) of the container, the location of
the valve-infra is bind-mounted into /app/valve-infra within the build
container.

However, if you wish to reprovision a running gateway, this location
won't exist, since we do not bundle the source code into the gateway
images. This commit allows the playbook runner to pass an HTTP(S) URL
for instance to a git tag in the case of production redeployments.

Credit to @craftyguy for the idea.