ci: Tag respective containers with latest
This is a bit of a tradeoff, we don't want silent upgrades (and hence
potentially silent breaks) but we also don't want to forget to ever
update things. The infra container is a particularly important one
here, ideally reboots will move to the latest and greatest, and that's
most convenient with a latest tag.

I added the same logic to the machine_registration container while I
was at it. Since after scripts are not working, you can't do this
trick with the ci-templates containers (the triggers), so for now they
remain as manual tags. Perhaps another useful thing to add to
ci-templates!