machine_registration: drop the VRAM size from the machine tags
Detection of the VRAM size is dependent on a vulkan driver, which in
turns requires a ton of dependencies which will be dropped in a future
commit.

An additional issue we found is that the way we detect the VRAM size is
not reliable, with one NAVI10 going from 7.5GiB to 2GB... We'll fix
this in a later series.