Makefile: add support for building the machine_registration container
This was the only container left that we needed to generate, while the
valve-infra-container was already best-generated using the Makefile.
This allows us to drop the local_build.sh file, since it is now
completely useless.

Funny how we went from a Makefile to local_build, then back to
Makefile :D