ansible: allow wg clients to access the public interface
This is quite useful as it allows accessing the
PDU/network router/... from the developers' machines.

Since we require information about the public interface to achieve
this, let's move the merge the wireguard mode into the network_setup
one... which is anyway where wireguard should have been to begin with!