executor/job: allow marking a machine as unfit for service
When the hardware of a machine changes, the machine registration
container will detect the difference in HW and prevent running on it.
However, the runs on this machine will keep failing until an admin
fixes the tags... which can lead to a relatively long down time.

To address this issue, this commit adds a console pattern in the
job description that will allow the user to specify one or multiple
console outputs that mean the machine needs to marked as unfit for
service, and do so when it is found. The job will then fail, and
Sergent Hartman can re-enroll the machine!