autologin: use the current console rather than tty1
TTY1 is already used by the console of b2c's initscript, this means
that when getty takes over it, the two get in control of
stdin/stdout/stderr. When we shut down the container, systemd stops
the getty on tty1, which closes stdin/stdout/stderr and get
b2c's initscript killed using SIGPIPE...

To avoid this issue and making running the valve infra container
easier to run on development machines, let's ask systemd to run our
dashboard on our current console!

Fixes: #49