move valve-infra-container-base build to new script
This moves the script for building the base container image to a new
script that is called by CI, and adds a new make target for it too.

This also changes BASE_IMAGE in the makefile to be settable via
environment at runtime, so you can specify an image in a local registry
for local debug/testing.