gitlab-ci/build_functions: don't push tags if branch is not default branch
This only allows pushing tags in CI if the branch for the pipeline is
the default branch (e.g. "master"), and disallowing pushing tags on
all other branches (for merge requests, etc)

Pushing tags when building locally, i.e. not in CI, is still allowed if
IMAGE_NAME_LATEST is set, which perserves the existing behavior.

fixes #54