ansible: remove dependencies definition for executor
Because these were listed here AND in the playbook, they were being
executed twice... Generally I prefer explicit dependencies listed like
this, since I was the one that added this config, but our playbook is
really only for 1 host and we manage the list of roles there so we can
tag them easily. So this is basically redundant.