build_functions: use skopeo rather than podman to tag latest
Skopeo has been written with the goal of copying containers between
different registries. This is exactly what we need when setting the
"latest" tag on a container.

By moving to skopeo, we remove the possibility that podman would not
have the image in its storage, and thus would fail to re-tag and push
the image back. Additionally, skopeo will not require you to download
the image first before pushing it... so let's just use it!